// packages/mermaid/src/rendering-util/types.ts
interface BaseNode {
  id: string;
  label?: string;
  description?: string[];
  parentId?: string;
  position?: string; // Keep, this is for notes 'left of', 'right of', etc. Move into nodeNode
  cssStyles?: string[]; // Renamed from `styles` to `cssStyles`
  cssCompiledStyles?: string[];
  cssClasses?: string; // Renamed from `classes` to `cssClasses`
  // style?: string; //REMOVE ✅
  // class?: string; //REMOVE ✅
  // labelText?: string; //REMOVE, use `label` instead  ✅
  // props?: Record<string, unknown>; //REMOVE  ✅
  // type: string; // REMOVE, replace with isGroup: boolean, default false  ✅
  // borders?: string; //REMOVE  ✅
  labelStyle?: string; // REMOVE - use cssStyles instead  ✅

  // Flowchart specific properties
  labelType?: string; // REMOVE? Always use markdown string, need to check for KaTeX - ⏳ wait with this one

  domId?: string; // When you create the node in the getData function you do not have the domId yet
  // Rendering specific properties for both Flowchart and State Diagram nodes
  dir?: string; // Only relevant for isGroup true, i.e. a sub-graph or composite state.
  haveCallback?: boolean;
  link?: string;
  linkTarget?: string;
  tooltip?: string;
  padding?: number; //REMOVE?, use from LayoutData.config - Keep, this could be shape specific
  isGroup?: boolean;
  width?: number;
  height?: number;
  // Specific properties for State Diagram nodes TODO remove and use generic properties
  intersect?: (point: any) => any;
  calcIntersect?: (bounds: Bounds, point: Point) => any;

  // Non-generic properties
  rx?: number; // Used for rounded corners in Rect, Ellipse, etc.Maybe it to specialized RectNode, EllipseNode, etc.
  ry?: number;

  useHtmlLabels?: boolean;
  centerLabel?: boolean; //keep for now.
  //Candidate for removal, maybe rely on labelStyle or a specific property labelPosition: Top, Center, Bottom

  //Node style properties
  backgroundColor?: string;
  borderColor?: string;
  borderStyle?: string;
  borderWidth?: number;
  labelTextColor?: string;
  labelPaddingX?: number;
  labelPaddingY?: number;

  // Flowchart specific properties
  x?: number;
  y?: number;

  look?: string;
  icon?: string;
  pos?: 't' | 'b';
  img?: string;
  assetWidth?: number;
  assetHeight?: number;
  defaultWidth?: number;
  imageAspectRatio?: number;
  constraint?: 'on' | 'off';
  layer?: number;
  order?: number;
  children?: NodeChildren;
  nodeId?: string;
  level?: number;
  descr?: string;
  type?: number;
  radius?: number;
  taper?: number;
  stroke?: string;
  colorIndex?: number;
}

/**
 * Group/cluster nodes, e.g. nodes that contain other nodes.
 */
export type NodeChildren = Node[];

export interface ClusterNode extends BaseNode {
  shape?: ClusterShapeID;
  isGroup: true;
  isEdgeLabel?: boolean;
  edgeStart?: string;
  edgeEnd?: string;
  isDummy?: boolean;
}
export interface NonClusterNode extends BaseNode {
  shape?: ShapeID;
  isGroup: false;
  isEdgeLabel?: boolean;
  edgeStart?: string;
  edgeEnd?: string;
  isDummy?: boolean;
}

// Common properties for any node in the system
export type Node = ClusterNode | NonClusterNode;
// ...
export interface RectOptions {
  rx: number;
  ry: number;
  labelPaddingX: number;
  labelPaddingY: number;
  classes: string;
}